# doc/api.yaml
openapi: 3.0.0
paths:
  /:
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - App
  /user:
    post:
      description: Creates a new user
      operationId: UsersController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: The user has been created.
        '400':
          description: Bad request. body does not contain required fields
      summary: Create user
      tags: &ref_0
        - User
    get:
      description: Gets all users
      operationId: UsersController_findAll
      parameters: []
      responses:
        '200':
          description: Successful operation
      summary: Get all users
      tags: *ref_0
  /user/{id}:
    get:
      description: Get single user by id
      operationId: UsersController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request. userId is invalid (not uuid)
        '404':
          description: User not found
      summary: Get single user by id
      tags: *ref_0
    put:
      description: Update library user information by UUID
      operationId: UsersController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordDto'
      responses:
        '200':
          description: The user has been updated.
        '400':
          description: Bad request. userId is invalid (not uuid)
        '403':
          description: oldPassword is wrong
        '404':
          description: User was not found.
      summary: Update user information
      tags: *ref_0
    delete:
      description: Delete user
      operationId: UsersController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Deleted successfully
        '400':
          description: Bad request. userId is invalid (not uuid)
        '404':
          description: User was not found.
      summary: Delete user
      tags: *ref_0
  /artist:
    post:
      description: Add new artist
      operationId: ArtistsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArtistDto'
      responses:
        '201':
          description: Artist is created
        '400':
          description: Bad request. body does not contain required fields
      summary: Add new artist
      tags: &ref_1
        - Artist
    get:
      description: Gets all artists
      operationId: ArtistsController_findAll
      parameters: []
      responses:
        '200':
          description: Successful operation
      summary: Get all artists
      tags: *ref_1
  /artist/{id}:
    get:
      description: Get single artist by id
      operationId: ArtistsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request. artistId is invalid (not uuid)
        '404':
          description: Artist not found
      summary: Get single artist by id
      tags: *ref_1
    put:
      description: Update library artist information by UUID
      operationId: ArtistsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArtistDto'
      responses:
        '200':
          description: The artist has been updated.
        '400':
          description: Bad request. artistId is invalid (not uuid)
        '404':
          description: Artist was not found.
      summary: Update artist information
      tags: *ref_1
    delete:
      description: Delete artist
      operationId: ArtistsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Deleted successfully
        '400':
          description: Bad request. artistId is invalid (not uuid)
        '404':
          description: Artist was not found.
      summary: Delete artist
      tags: *ref_1
  /album:
    post:
      description: Add new album information
      operationId: AlbumsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlbumDto'
      responses:
        '201':
          description: Album is created
        '400':
          description: Bad request. body does not contain required fields
      summary: Add new album
      tags: &ref_2
        - Album
    get:
      description: Gets all library albums list
      operationId: AlbumsController_findAll
      parameters: []
      responses:
        '200':
          description: Successful operation
      summary: Get albums list
      tags: *ref_2
  /album/{id}:
    get:
      operationId: AlbumsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request. albumId is invalid (not uuid)
        '404':
          description: Album not found
      summary: Get albums list
      tags: *ref_2
    put:
      description: Update library album information by UUID
      operationId: AlbumsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlbumDto'
      responses:
        '200':
          description: The album has been updated.
        '400':
          description: Bad request. albumId is invalid (not uuid)
        '404':
          description: Album was not found.
      summary: Update album information
      tags: *ref_2
    delete:
      description: Delete album from library
      operationId: AlbumsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Deleted successfully
        '400':
          description: Bad request. albumId is invalid (not uuid)
        '404':
          description: Album was not found.
      summary: Delete album
      tags: *ref_2
  /track:
    post:
      description: Add new track information
      operationId: TracksController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrackDto'
      responses:
        '201':
          description: The track has been created.
        '400':
          description: Bad request. body does not contain required fields
      summary: Add new track
      tags: &ref_3
        - Track
    get:
      description: Gets all library tracks list
      operationId: TracksController_findAll
      parameters: []
      responses:
        '200':
          description: Successful operation
      summary: Get tracks list
      tags: *ref_3
  /track/{id}:
    get:
      description: Get single track by id
      operationId: TracksController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request. trackId is invalid (not uuid)
        '404':
          description: Track not found
      summary: Get single track by id
      tags: *ref_3
    put:
      description: Update library track information by UUID
      operationId: TracksController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTrackDto'
      responses:
        '200':
          description: The track has been updated.
        '400':
          description: Bad request. trackId is invalid (not uuid)
        '404':
          description: Track was not found.
      summary: Update track information
      tags: *ref_3
    delete:
      description: Delete track
      operationId: TracksController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Deleted successfully
        '400':
          description: Bad request. trackId is invalid (not uuid)
        '404':
          description: Track was not found.
      summary: Delete track
      tags: *ref_3
info:
  title: REST Service
  description: >-
    Home Library Service. Users can create, read, update, delete data about
    Artists, Tracks and Albums, add them to Favorites in their own Home Library.
  version: '1.0'
  contact: {}
tags:
  - name: Home Library Service
    description: ''
servers: []
components:
  schemas:
    CreateUserDto:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
      required:
        - login
        - password
    UpdatePasswordDto:
      type: object
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
      required:
        - oldPassword
        - newPassword
    CreateArtistDto:
      type: object
      properties:
        name:
          type: string
        grammy:
          type: boolean
      required:
        - name
        - grammy
    UpdateArtistDto:
      type: object
      properties:
        name:
          type: string
        grammy:
          type: boolean
    CreateAlbumDto:
      type: object
      properties:
        name:
          type: string
        year:
          type: number
        artistId:
          type: string
      required:
        - name
        - year
        - artistId
    UpdateAlbumDto:
      type: object
      properties:
        name:
          type: string
        year:
          type: number
        artistId:
          type: string
    CreateTrackDto:
      type: object
      properties:
        name:
          type: string
        artistId:
          type: string
        albumId:
          type: string
        duration:
          type: number
      required:
        - name
        - artistId
        - albumId
        - duration
    UpdateTrackDto:
      type: object
      properties:
        name:
          type: string
        artistId:
          type: string
        albumId:
          type: string
        duration:
          type: number